+++ ts/object/src/lib/decorators/build.ts
@@ -1,0 +1,24 @@
+import { meta } from '../meta'
+
+
+export function build( method?:string|Object, ...args ):any {
+
+    function build( target:any, name: string,  descriptor?: TypedPropertyDescriptor<Function>) {
+        if ( descriptor ) throw new Error("Cannot use the build decorator on a method") 
+        if ( method === undefined ) method = `_build_${name}`
+        meta(target).property(name).default( o => o[<string>method].call(o,o) )
+    }
+
+    if ( method instanceof Object ) {
+        let [target, name, descriptor] = [method, ...args]
+
+        method = undefined
+        
+        return build(target, name, <any>descriptor)
+    }
+
+    else {
+        return build
+    }
+
+}