+++ ts/object/CHANGES.md
@@ -1,2 +1,11 @@
 
 
+v 1.0.1
+- @lazy can now be called without arguments
+- added @build decorator
+- added deflate function
+
+v 0.1.3
+- initial
+
+